chore: 同步test & example
