+++ examples/heatmap/basic/index.zh.md
@@ -1,4 +1,4 @@
 ---
-title: 基础热力图
+title: 热力图
 order: 0
 ---
