+++ examples/heatmap/density/index.en.md
@@ -1,6 +1,6 @@
 ---
-title: Matrix Chart
-order: 1
+title: DensityHeatmap Chart
+order: 0
 ---
 
 Description about this component.
