+++ examples/heatmap/density/index.zh.md
@@ -1,0 +1,4 @@
+---
+title: 密度热力图
+order: 1
+---
