--- examples/heatmap/matrix/index.zh.md
@@ -1,4 +1,0 @@
----
-title: 色块热力图
-order: 1
----
